requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
python-dotenv
